requirements.txt
aiogram>=3.*
vk_api>=11.9.9
python-dotenv==1.0.1
